\name{amc12}
\docType{data}
\alias{amc12}
\title{High Math Achievement at the Amc 12 Contest}
\description{a cross-section of 1984 schools

    \emph{country} :  United States 

    \emph{package} : count

    \emph{JEL codes: }H75, I21, I24, I28, R23

}
\usage{data(amc12)}
\format{A dataframe containing:
  \describe{
    \item{over100}{number of students with a score greater or equal to 100 to the AMC12 test}
    \item{over120}{number of students with a score greater or equal to 120 to the AMC12 test}
    \item{students}{number of students in the school}
    \item{bachelors}{percentage of bachelors}
    \item{graduates}{percentage of graduates}
    \item{asian}{percentage of asian}
    \item{black}{percentage of blacks}
    \item{hispanic}{percentage of hispanics}
    \item{medinc}{median income}
    \item{freelunch}{percentage of free lunch}
    \item{t1}{title 1 school}
    \item{urban}{percentage of urbans}
    \item{female}{percentage of females}
    \item{satact}{SAT/ACT participation}
  }
}
\source{American Economic Association Data Archive : \url{https://www.aeaweb.org/aer/}}
\references{
Ellison, Glenn and Ashley Swanson (2016) \dQuote{Do Schools Matter for High Math Achievement ? Evidence From the American Mathematics Competitions}, \emph{American Economic Review}, \bold{106(6)}, 1244-1277, \doi{10.1257/aer.20140308}
.

}
\keyword{datasets}
