\name{asymptotic}
\docType{data}
\alias{asymptotic}
\title{Asymptotic Theory and Econometric Practise}
\description{a cross-section of 733 articles from 1970 to 1980

    \emph{country} :  world 

    \emph{package} : count

    \emph{JEL codes: }C01, C87

}
\usage{data(asymptotic)}
\format{A dataframe containing:
  \describe{
    \item{id}{article id}
    \item{neq}{number of equations}
    \item{nreg}{number of regressors}
    \item{nobs}{number of observations}
    \item{author}{article's author}
    \item{journal}{journal where the article has been published}
    \item{year}{year}
    \item{page}{number of pages.}
    \item{subject}{subject of the article}
    \item{collection}{yes or no}
  }
}
\source{Journal of Applied Econometrics Data Archive : \url{http://qed.econ.queensu.ca/jae/}}
\references{
Koenker, Roger (1988) \dQuote{Asymptotic Theory and Econometric Practise}, \emph{Journal of Applied Econometrics}, \bold{3}, 139-147, \doi{10.1002/jae.3950030205}
.

Koenker, Roger and Achim Zeileis (2009) \dQuote{On Reproducible Econometric Research}, \emph{Journal of Applied Econometrics}, \bold{25}, 833-847, \doi{10.1002/jae.1083}
.

}
\keyword{datasets}
