\name{bids}
\docType{data}
\alias{bids}
\title{Bids Received By Us Firms}
\description{a cross-section of 126 firm

    \emph{country} :  United States 

    \emph{package} : 

}
\usage{data(bids)}
\format{A dataframe containing:
  \describe{
    \item{numbids}{number of bids}
    \item{legrest}{legal restructuring}
    \item{realrest}{real restructuring}
    \item{finrest}{financial restructuring}
    \item{whtknght}{white knight}
    \item{bidprem}{bid premium}
    \item{insthold}{institutional holdings}
    \item{size}{size measured in billions}
    \item{regulation}{regulation}
  }
}
\source{Journal of Applied Econometrics Data Archive : \url{http://qed.econ.queensu.ca/jae/}}
\references{
Cameron, A. Colin and Per Johansson (1997) \dQuote{Count Data Regression Using Series Expansion: with Applications}, \emph{Journal of Applied Econometrics}, \bold{12(3)}, 203-224.

}
\keyword{datasets}
