\name{cigmales}
\docType{data}
\alias{cigmales}
\title{Cigarette Smoking Behaviour}
\description{a cross-section of 6160 individuals from 1979 to 1980

    \emph{country} :  United States 

    \emph{package} : 

}
\usage{data(cigmales)}
\format{A dataframe containing:
  \describe{
    \item{cigarettes}{number of daily cigarettes smoked}
    \item{habit}{smoking habit stock measure}
    \item{price}{state-level average per-pack price of cigarettes in 1979}
    \item{restaurant}{an indicator of whether the individual's state of residence had restrictions on smoking in restaurants in place in 1979}
    \item{income}{family income in thousands}
    \item{age}{age in years}
    \item{educ}{schooling in years}
    \item{famsize}{number of family members}
    \item{race}{a factor with levels "other" and "white"}
    \item{reslgth}{number of years the state's restaurant smoking restrictions had been in place in 1979}
    \item{lagprice}{one-year lag of cigarette price}
  }
}
\source{kindly provided by John Mullahy}
\references{
Mullahy, John (1997) \dQuote{Instrumental-variable Estimation of Count Data Models: Applications to Models of Cigarette Smoking Behavior}, \emph{Review of Economics and Statistics}, \bold{79(4)}, 586-593.

}
\keyword{datasets}
