\name{doctor_ger}
\docType{data}
\alias{doctor_ger}
\title{Doctor Visits in Germany}
\description{a cross-section of 5096 individuals from 1985

    \emph{country} :  Germany 

    \emph{package} : 

    \emph{JEL codes: }C12, C13, C25, I10

}
\usage{data(doctor_ger)}
\format{A dataframe containing:
  \describe{
    \item{gdoctor}{number of visits to a general practinioner in the last quarter (before time of survey),}
    \item{sdoctor}{number of visits to a specialist in the last quarter (before time of survey) with the exception of visits to a gynaecologist, pediatrician or dentist,}
    \item{female}{a dummy for females,}
    \item{single}{a dummy for single,}
    \item{age}{age in years}
    \item{income}{monthly income, net of income taxes and social insurance rates; including regular monthly subsidies such as housing allowances and other transfers}
    \item{chronic}{a dummy for chronic illness or complaints for at least one year,}
    \item{privins}{a dummy for private medical insurance in the last year}
    \item{school}{number of years of education beyond the age of sixteen}
    \item{heavylab}{a dummy; person has a position where heavy physical work is required,}
    \item{stress}{a dummy; position involves a high level of stress}
    \item{variety}{job offers a lot of variety}
    \item{selfdet}{a dummy; individual can plan and carry out his job tasks}
    \item{control}{a dummy; work performance is strictly controlled}
    \item{physdens}{number of licensed physicians per 100,000 residents in the place of residence}
    \item{unemp}{duration of unemployment in months in the previous year}
    \item{hospm7}{a dummy; more than seven days hospitalized in previous year}
    \item{sickm14}{a dummy; missed more than 14 work days because of sickness in previous year}
    \item{disability}{a dummy; degree of disability is greater than 20 percent}
    \item{pop}{population of the place of residence, a factor with levels `">100k"`, `"<5k"`, `"5-20k"` and `"20k-100k"`.}
  }
}
\source{kindly provided by Winfried Pohlmeier}
\references{
Pohlmeier, Winfried and Volker Ulrich (1995) \dQuote{An Econometric Model of the Two-part Decisionmaking Process in the Demand for Health Care}, \emph{Journal of Human Ressources}, \bold{30(2)}, 339-361.

Santos Silva, Joao M.C. and Frank Windmeijer (2001) \dQuote{Two-part Multiple Spell Models for Health Care Demand}, \emph{Journal of Econometrics}, \bold{104}, 67-89.

}
\keyword{datasets}
