\name{doctor_us}
\docType{data}
\alias{doctor_us}
\title{Health Care Utilization in the United States}
\description{a cross-section of 36111 individuals from 1992

    \emph{country} :  United States 

    \emph{package} : 

}
\usage{data(doctor_us)}
\format{A dataframe containing:
  \describe{
    \item{age}{age in years}
    \item{educ}{years of education}
    \item{poverty}{NA}
    \item{famsize}{family size}
    \item{doctvisits}{number of doctor visits in the last 12 months}
    \item{male}{a dummy}
    \item{white}{a dummy}
    \item{married}{a dummy}
    \item{working}{a dummy}
    \item{school}{a dummy}
    \item{health}{a factor with levels `poor`, `fair`, `good`, `verygood` and `excellent`}
  }
}
\source{JAE data archive}
\references{
Mroz, Thomas A. (2012) \dQuote{A Simple, Flexible Estimator for Count and Other Ordered Discrete Data}, \emph{Journal of Applied Econometrics}, \bold{27(4)}, 646-665.

Mullahy, J. (1998) \dQuote{Much Ado About Two: Reconsidering Retransformation and the Two-part Model in Health Econometrics}, \emph{Journal of Health Economics}, \bold{17}, 247-281.

}
\keyword{datasets}
