\name{drinks}
\docType{data}
\alias{drinks}
\title{Physician Advice on Alcohol Consumption}
\description{a cross-section of 2467 individuals from 1990

    \emph{country} :  United states 

    \emph{package} : 

}
\usage{data(drinks)}
\format{A dataframe containing:
  \describe{
    \item{drinks}{number of drinks in the past 2 weeks}
    \item{advice}{1 if reveived a drining advice}
    \item{age}{age in 10 years cathegories}
    \item{race}{a factor with levels `white`, `black` and `other`}
    \item{marital}{marital status, one of `single`, `married`, `widow`, `separated`}
    \item{region}{one of `west`, `northeast`, `midwest` and `south`}
    \item{empstatus}{one of `other`, `emp` and `unemp`}
    \item{limits}{limits on daily activities, one of `none`, `some` and `major`}
    \item{income}{monthly income ($1000)}
    \item{educ}{education in years}
    \item{medicare}{insurance through medicare}
    \item{medicaid}{insurance through medicaid}
    \item{champus}{military insurance}
    \item{hlthins}{health insurance}
    \item{regmed}{regoular source of care}
    \item{dri}{see same doctor}
    \item{diabete}{have diabetes}
    \item{hearthcond}{have heart condition}
    \item{stroke}{have stroke}
  }
}
\source{JAE data archive}
\references{
Kenkel, Donald S. and Joseph V. Terza (2001) \dQuote{The Effect of Physician Advice on Alcohol Consumption: Count Regression with An Endogenous Treatment Effect}, \emph{Journal of Applied Econometrics}, \bold{16(2)}, 165-184.

}
\keyword{datasets}
