\name{elderly}
\docType{data}
\alias{elderly}
\title{Demand for Medical Care By the Elderly}
\description{a cross-section of 4406 individuals from 1987

    \emph{country} :  United States 

    \emph{package} : 

}
\usage{data(elderly)}
\format{A dataframe containing:
  \describe{
    \item{ofp}{number of physician office visits}
    \item{ofnp}{number of non-physician office visits}
    \item{opp}{number of physician outpatient visits}
    \item{opnp}{number of non-physician outpatient visits}
    \item{emr}{number of emergency room visits}
    \item{hosp}{number hospitalizations}
    \item{health}{a factor with levels `medium`, `excellent` and `poor`}
    \item{numchron}{number of chronic conditions}
    \item{adldiff}{a dummy :the person has a condition that limits activities of daily living}
    \item{region}{a factor with levels `south`, `noreast`, `midwest` and `west`}
    \item{age}{age in years (divided by 10)}
    \item{race}{a factor with levels `other` and `black`}
    \item{sex}{a factor with levels `female` and `male`}
    \item{marital}{a factor with levels `other` and `married`}
    \item{school}{number of years of education}
    \item{income}{family income in $10,000}
    \item{emp}{is the person employed ? A factor with levels `no` and `yes`}
    \item{insurance}{insurance, a factor with levels `other`, `privins` and `medicaid`}
  }
}
\source{Journal of Applied Econometrics Data Archive : \url{http://qed.econ.queensu.ca/jae/}}
\references{
Deb, Partha and Parvin K. Trivedi (1997) \dQuote{Demand for Medical Care By the Elderly: A Finite Mixture Approach}, \emph{Journal of Applied Econometrics}, \bold{12(3)}, 313-336.

}
\keyword{datasets}
