\name{federiv}
\docType{data}
\alias{federiv}
\title{Foreign Exchange Derivatives Use By Large Us Bank Holding Companies}
\description{a cross-section of 794 banks from 1996 to 2000

    \emph{country} :  United States 

    \emph{package} : 

}
\usage{data(federiv)}
\format{A dataframe containing:
  \describe{
    \item{federiv}{foreign exchange derivatives use, a dummy}
    \item{optval}{option awards}
    \item{eqrat}{leverage}
    \item{bonus}{bonus}
    \item{ltass}{logarithm of total assets}
    \item{linsown}{logarithm of the percentage of the total shares outstanding that are owned by officers and directors}
    \item{linstown}{logarithm of the percentage of the total shares outstanding that are owned by all institutional investors}
    \item{roe}{return on equity}
    \item{mktbk}{market to book ratio}
    \item{perfor}{foreign to total interest income ratio}
    \item{dealdum}{derivative dealer activity dummy}
    \item{div}{dividends paid}
    \item{year}{year, from 1996 to 2000}
    \item{no_emp}{number of employees}
    \item{no_subs}{number of subsidiaries}
    \item{no_off}{number of offices}
    \item{ceo_age}{CEO age}
    \item{gap}{12 month maturity mismatch}
    \item{cfa}{ratio of cash flow to total assets}
  }
}
\source{Lee Adkin's home page https://learneconometrics.com/}
\references{
Adkins, Lee C. (2012) \dQuote{Testing Parameter Significance in Instrumental Variables Probit Estimators: Some Simulation}, \emph{Journal of Statistical Computation and Simulation}, \bold{83(10)}, 1415-1436.

Carter, David; Simpson, Gary and Lee Adkins (2007) \dQuote{Managerial Incentives and the Use of Foreign-exchange Derivatives By Banks}, \emph{The Journal of Financial Research}, \bold{20(3)}, 399-413.

}
\keyword{datasets}
