\name{health_reform}
\docType{data}
\alias{health_reform}
\title{Health Care Reform in Germany}
\description{a cross-section of 9197 individuals from 1995 to 1999

    \emph{country} :  Germany 

    \emph{package} : 

    \emph{JEL codes: }I11, I18

}
\usage{data(health_reform)}
\format{A dataframe containing:
  \describe{
    \item{id}{individual index}
    \item{year}{time index}
    \item{visits}{number of doctor visits}
    \item{age}{age}
    \item{male}{1 for males}
    \item{educ}{years of education}
    \item{married}{1 for married}
    \item{hsize}{household size}
    \item{sport}{1 if involved in a sport}
    \item{sozh}{1 if receive welfare paiement}
    \item{quarter}{quarter of the interview}
    \item{health}{self-assesment of the health status (one of `'medium'`, `'good`' and `'bad'`)}
    \item{empl}{employment status, one of `'inactive'`, '`unemp`', '`ptime`' (part-time job), '`ftime`' (full-time job)}
    \item{income}{monthly gross income}
  }
}
\source{Journal of Applied Econometrics Data Archive : \url{http://qed.econ.queensu.ca/jae/}}
\references{
Winkelmann, Rainer (2004) \dQuote{Health Care Reform and the Number of Doctor Visits: An Econometric Analysis}, \emph{Journal of Applied Econometrics}, \bold{19(4)}, 455-472.

}
\keyword{datasets}
