\name{hospitalization}
\docType{data}
\alias{hospitalization}
\title{Economic Incentives and Hospitalization in Germany}
\description{yearly observations of 5180  from 1984 to 1994

    \emph{number of observations} : 30590

    \emph{number of  time-series } :  NA 

    \emph{country} :  

    \emph{package} : 

}
\usage{data(hospitalization)}
\format{A dataframe containing:
  \describe{
    \item{hospital}{number of yearly inpatient stays in hospital}
    \item{age}{age of the individual in years}
    \item{gender}{a factor with levels `female` and `male`}
    \item{private}{a dummy for full coverage by private health insurance}
    \item{copayment}{a dummy for full coverage by private health insurance with copayment obligation}
    \item{public}{a dummy for public, i.e. non-private insurance scheme including those insured voluntarily and as family members}
    \item{voluntary}{a dummy for voluntarily in public insurance scheme}
    \item{family}{a dummy for public insurance scheme for non-working spouse and children of paying members of public insurances}
    \item{aok}{a dummy for public insurance in Allgemeine Ortskrankenkasse, the insurance company legally obliged to accept all risks}
    \item{additional}{a dummy for public insurance with voluntary additional coverage through a private scheme}
    \item{chronic}{a dummy for chronic conditions}
    \item{handicap}{a dummy for handicapped}
    \item{income}{monthly net income of the household (DM)}
    \item{distance}{a dummy for place of living is outside of a city centre}
    \item{married}{a dummy for married}
    \item{secondary}{a dummy for educational level at least secondary school (Realschule)}
    \item{university}{a dummy for degree from a university or technical college}
    \item{apprenticeship}{a dummy for passed vocational training}
    \item{healthjob}{a dummy for working in a health related}
    \item{inlabour}{a dummy for in the labour force, i.e. working or unemployed}
    \item{bluecol}{a dummy for blue collar worker}
    \item{whitecol}{a dummy for white collar worker}
    \item{civil}{a dummy for civil servant}
    \item{selfemp}{a dummy for self-employed}
    \item{ptjob}{a dummy for part-time job}
    \item{western}{a dummy for EU, US, Canadian, or Swiss national}
    \item{nationelse}{a dummy for other non-German nationals}
    \item{children}{a dummy for children below age 16 in the household}
    \item{index}{individual index}
  }
}
\source{Journal of Applied Econometrics Data Archive : \url{http://qed.econ.queensu.ca/jae/}}
\references{
Geil, Peter; Million, Andreas; Rotte, Ralf and Klaus F. Zimmermann (1997) \dQuote{Economic Incentives and Hospitalization in Germany}, \emph{Journal of Applied Econometrics}, \bold{12(3)}, 295-312.

}
\keyword{datasets}
