\name{Index.economics}
\docType{data}
\alias{Index.Economic Topics}
\title{Economic Topics}
\description{
\itemize{
\item consumer behavior  
\itemize{
\item \code{\link{doctor_hospital}} : Doctor and hospital visits
}
\item consumer behaviour  
\itemize{
\item \code{\link{somerville}} : Visits to lake Somerville
}
\item economics of crime  
\itemize{
\item \code{\link{parking}} : Parking violations at the United Nations
\item \code{\link{urban_crimes}} : Urban Crimes in Pitsburg
}
\item Education  
\itemize{
\item \code{\link{amc12}} : High math achievement at the AMC 12 contest
}
\item finance  
\itemize{
\item \code{\link{federiv}} : Foreign exchange derivatives use by large US bank holding companies
}
\item health econometrics  
\itemize{
\item \code{\link{doctor_ger}} : Doctor visits in Germany
}
\item health economics  
\itemize{
\item \code{\link{birthwt}} : Cigarette smoking and birth weight
\item \code{\link{cigmales}} : Cigarette smoking behaviour
\item \code{\link{doctor_aus}} : Doctor visits in Australia
\item \code{\link{doctor_cal}} : Doctor Visits in California
\item \code{\link{doctor_us}} : Health care utilization in the United States
\item \code{\link{drinks}} : Physician advice on alcohol consumption
\item \code{\link{health_reform}} : Health care reform in Germany
\item \code{\link{hospitalization}} : Economic incentives and hospitalization in Germany
}
\item Health Economics  
\itemize{
\item \code{\link{elderly}} : Demand for medical care by the elderly
\item \code{\link{health_ins}} : Health Insurance and Doctor Visits
}
\item industrial economics  
\itemize{
\item \code{\link{cartels}} : Leniency and Cartels Enforcement
\item \code{\link{innovation}} : Institutional Ownership and Innovation
}
\item industrial organization  
\itemize{
\item \code{\link{bids}} : Bids received by US firms
}
\item labor economics  
\itemize{
\item \code{\link{jobmob}} : Job mobility
\item \code{\link{publications}} : Doctoral publications
\item \code{\link{strikes}} : strikes frequency and duration
}
\item microeconomics  
\itemize{
\item \code{\link{majordrg}} : Credit Card and Derogatory
}
\item producer behavior  
\itemize{
\item \code{\link{PatentsRD}} : Patents, R&D and technological spillovers for a panel of firms
}
\item theory  
\itemize{
\item \code{\link{asymptotic}} : Asymptotic theory and econometric practise
}
\item transport economics  
\itemize{
\item \code{\link{trips}} : Determinants of household trip taking
}
}
}
\keyword{documentation}
