\name{Index.jel}
\docType{data}
\alias{Index.JEL codes}
\title{JEL codes}
\description{
\itemize{
\item \bold{ 2120 } :  NA 
\itemize{
\item \code{\link{strikes}} : strikes frequency and duration
}
\item \bold{ 35 } :  NA 
\itemize{
\item \code{\link{somerville}} : Visits to lake Somerville
}
\item \bold{ 6310 } :  NA 
\itemize{
\item \code{\link{strikes}} : strikes frequency and duration
}
\item \bold{ 8321 } :  NA 
\itemize{
\item \code{\link{strikes}} : strikes frequency and duration
}
\item \bold{ C01 } :  Econometrics 
\itemize{
\item \code{\link{asymptotic}} : Asymptotic theory and econometric practise
}
\item \bold{ C12 } :  Hypothesis Testing: General 
\itemize{
\item \code{\link{doctor_ger}} : Doctor visits in Germany
}
\item \bold{ C13 } :  Estimation: General 
\itemize{
\item \code{\link{doctor_ger}} : Doctor visits in Germany
}
\item \bold{ C25 } :  Single Equation Models; Single Variables: Discrete Regression and Qualitative Choice Models; Discrete Regressors; Proportions; Probabilities 
\itemize{
\item \code{\link{doctor_cal}} : Doctor Visits in California
\item \code{\link{doctor_ger}} : Doctor visits in Germany
}
\item \bold{ C51 } :  Model Construction and Estimation 
\itemize{
\item \code{\link{doctor_cal}} : Doctor Visits in California
\item \code{\link{somerville}} : Visits to lake Somerville
}
\item \bold{ C87 } :  Econometric Software 
\itemize{
\item \code{\link{asymptotic}} : Asymptotic theory and econometric practise
}
\item \bold{ D12 } :  Consumer Economics: Empirical Analysis 
\itemize{
\item \code{\link{health_ins}} : Health Insurance and Doctor Visits
}
\item \bold{ D43 } :  Market Structure, Pricing, and Design: Oligopoly and Other Forms of Market Imperfection 
\itemize{
\item \code{\link{cartels}} : Leniency and Cartels Enforcement
}
\item \bold{ D61 } :  Allocative Efficiency; Cost-Benefit Analysis 
\itemize{
\item \code{\link{somerville}} : Visits to lake Somerville
}
\item \bold{ G22 } :  Insurance; Insurance Companies; Actuarial Studies 
\itemize{
\item \code{\link{health_ins}} : Health Insurance and Doctor Visits
}
\item \bold{ G23 } :  Pension Funds; Non-bank Financial Institutions; Financial Instruments; Institutional Investors 
\itemize{
\item \code{\link{innovation}} : Institutional Ownership and Innovation
}
\item \bold{ G32 } :  Financing Policy; Financial Risk and Risk Management; Capital and Ownership Structure; Value of Firms; Goodwill 
\itemize{
\item \code{\link{innovation}} : Institutional Ownership and Innovation
}
\item \bold{ H75 } :  State and Local Government: Health; Education; Welfare; Public Pensions 
\itemize{
\item \code{\link{amc12}} : High math achievement at the AMC 12 contest
}
\item \bold{ I10 } :  Health: General 
\itemize{
\item \code{\link{doctor_ger}} : Doctor visits in Germany
}
\item \bold{ I11 } :  Analysis of Health Care Markets 
\itemize{
\item \code{\link{doctor_cal}} : Doctor Visits in California
\item \code{\link{doctor_hospital}} : Doctor and hospital visits
\item \code{\link{health_ins}} : Health Insurance and Doctor Visits
\item \code{\link{health_reform}} : Health care reform in Germany
}
\item \bold{ I12 } :  Health Behavior 
\itemize{
\item \code{\link{health_ins}} : Health Insurance and Doctor Visits
}
\item \bold{ I13 } :  Health Insurance, Public and Private 
\itemize{
\item \code{\link{health_ins}} : Health Insurance and Doctor Visits
}
\item \bold{ I18 } :  Health: Government Policy; Regulation; Public Health 
\itemize{
\item \code{\link{health_ins}} : Health Insurance and Doctor Visits
\item \code{\link{health_reform}} : Health care reform in Germany
}
\item \bold{ I21 } :  Analysis of Education 
\itemize{
\item \code{\link{amc12}} : High math achievement at the AMC 12 contest
}
\item \bold{ I24 } :  Education and Inequality 
\itemize{
\item \code{\link{amc12}} : High math achievement at the AMC 12 contest
}
\item \bold{ I28 } :  Education: Government Policy 
\itemize{
\item \code{\link{amc12}} : High math achievement at the AMC 12 contest
}
\item \bold{ I38 } :  Welfare, Well-Being, and Poverty: Government Programs; Provision and Effects of Welfare Programs 
\itemize{
\item \code{\link{doctor_cal}} : Doctor Visits in California
}
\item \bold{ K21 } :  Antitrust Law 
\itemize{
\item \code{\link{cartels}} : Leniency and Cartels Enforcement
}
\item \bold{ K42 } :  Illegal Behavior and the Enforcement of Law 
\itemize{
\item \code{\link{parking}} : Parking violations at the United Nations
}
\item \bold{ L12 } :  Monopoly; Monopolization Strategies 
\itemize{
\item \code{\link{cartels}} : Leniency and Cartels Enforcement
}
\item \bold{ L13 } :  Oligopoly and Other Imperfect Markets 
\itemize{
\item \code{\link{cartels}} : Leniency and Cartels Enforcement
}
\item \bold{ L25 } :  Firm Performance: Size, Diversification, and Scope 
\itemize{
\item \code{\link{innovation}} : Institutional Ownership and Innovation
}
\item \bold{ L60 } :  Industry Studies: Manufacturing: General 
\itemize{
\item \code{\link{PatentsRD}} : Patents, R&D and technological spillovers for a panel of firms
}
\item \bold{ L83 } :  Sports; Gambling; Restaurants; Recreation; Tourism 
\itemize{
\item \code{\link{somerville}} : Visits to lake Somerville
}
\item \bold{ M10 } :  Business Administration: General 
\itemize{
\item \code{\link{innovation}} : Institutional Ownership and Innovation
}
\item \bold{ O30 } :  Innovation; Research and Development; Technological Change; Intellectual Property Rights: General 
\itemize{
\item \code{\link{PatentsRD}} : Patents, R&D and technological spillovers for a panel of firms
}
\item \bold{ O31 } :  Innovation and Invention: Processes and Incentives 
\itemize{
\item \code{\link{innovation}} : Institutional Ownership and Innovation
}
\item \bold{ O34 } :  Intellectual Property and Intellectual Capital 
\itemize{
\item \code{\link{innovation}} : Institutional Ownership and Innovation
}
\item \bold{ Q26 } :  Recreational Aspects of Natural Resources 
\itemize{
\item \code{\link{somerville}} : Visits to lake Somerville
}
\item \bold{ R23 } :  Urban, Rural, Regional, Real Estate, and Transportation Economics: Regional Migration; Regional Labor Markets; Population; Neighborhood Characteristics 
\itemize{
\item \code{\link{amc12}} : High math achievement at the AMC 12 contest
\item \code{\link{health_ins}} : Health Insurance and Doctor Visits
}
\item \bold{ Z13 } :  Economic Sociology; Economic Anthropology; Language; Social and Economic Stratification 
\itemize{
\item \code{\link{parking}} : Parking violations at the United Nations
}
}
}
\keyword{documentation}
