\name{jobmob}
\docType{data}
\alias{jobmob}
\title{Job Mobility}
\description{a cross-section of 1962 individuals

    \emph{country} :  Germany 

    \emph{package} : count

}
\usage{data(jobmob)}
\format{A dataframe containing:
  \describe{
    \item{jobch}{number of job changes}
    \item{unemp}{number of unemployment spells}
    \item{educ}{education}
    \item{exper}{experience}
    \item{union}{union membership}
    \item{marital}{marital status (married or single)}
    \item{nation}{nationality (foreigner or German)}
    \item{qualif}{qualification (one of ordinary / qualified blue / white collar)}
  }
}
\source{kindly provided by Rainer Winkelmann}
\references{
Winkelman, Rainer(2008) Econometric analysis of count data, Springer.

}
\keyword{datasets}
