\name{majordrg}
\docType{data}
\alias{majordrg}
\title{Credit Card and Derogatory}
\description{a cross-section of 1319 individuals from 1988

    \emph{country} :  United States 

    \emph{package} : 

}
\usage{data(majordrg)}
\format{A dataframe containing:
  \describe{
    \item{cardhldr}{1 for cardholders, 0 for denied applicants}
    \item{majordrg}{count of major derogatory reports (long delinquincies) from credit bureau}
    \item{age}{age in years and thwelfhs of a year}
    \item{income}{primary and additional income}
    \item{expinc}{average monthly expenditure divided by monthly income}
    \item{avgexp}{average monthly expense}
    \item{ownrent}{1 if own home, 0 if rent}
    \item{selfemp}{1 if self employed, 0 otherwise}
    \item{depndt}{dependents, missing data converted to 1}
    \item{incper}{income per family member}
    \item{curadd}{months of current adress}
    \item{major}{}
    \item{active}{}
  }
}
\source{GREE:03}
\references{
}
\keyword{datasets}
