\name{parking}
\docType{data}
\alias{parking}
\title{Parking Violations at the United Nations}
\description{a pseudo-panel of 149 countries

    \emph{number of observations} : 298

    \emph{number of  individual observations } :  2 

    \emph{country} :  world 

    \emph{package} : count

    \emph{JEL codes: }K42, Z13

}
\usage{data(parking)}
\format{A dataframe containing:
  \describe{
    \item{violations}{number of parking violations commited by diplomats of a country}
    \item{corruption}{corruption's index for the country}
    \item{post}{a factor with level equal to \code{yes} for the post-legal enforcment period}
    \item{staff}{the number of diplomats of the country}
    \item{gdp}{the per capita gdp of the country}
    \item{region}{the region of the world}
    \item{country}{country acronym}
  }
}
\source{author's website \url{http://emiguel.econ.berkeley.edu/}}
\references{
Fisman, Raymond and Edward Miguel (2007) \dQuote{Corruption, Norms, and Legal Enforcement: Evidence From Diplomatic Parking Tickets}, \emph{Journal of Political Economy}, \bold{115(6)}, 1020-1048, \doi{10.1086/527495}
.

}
\keyword{datasets}
