\name{publications}
\docType{data}
\alias{publications}
\title{Doctoral Publications}
\description{a cross-section of 915 individuals

    \emph{country} :  United States 

    \emph{package} : 

}
\usage{data(publications)}
\format{A dataframe containing:
  \describe{
    \item{articles}{number of doctoral publications in the final three years of Ph.D. studies}
    \item{sex}{sex}
    \item{marital}{marital status}
    \item{kids}{number of young kids}
    \item{phd}{prestige of the Ph.D. department}
    \item{mentor}{number of articles of mentor in the previous three years}
  }
}
\source{https://stats.oarc.ucla.edu/other/examples/cldv/}
\references{
Long, John Scott(1997) Regression models for categorical and limited dependent variables, Sage Publications.

}
\keyword{datasets}
