\name{somerville}
\docType{data}
\alias{somerville}
\title{Visits to Lake Somerville}
\description{a cross-section of 659 individuals from 1980

    \emph{country} :  United States 

    \emph{package} : count

    \emph{JEL codes: }Q26, D61, C51, 35, L83

}
\usage{data(somerville)}
\format{A dataframe containing:
  \describe{
    \item{visits}{number of boating trips to Lake Somerville, East Texas, in 1980}
    \item{quality}{facility's subjective quality ranking}
    \item{ski}{respondent's taste for water-skiing}
    \item{income}{income, categorical variable}
    \item{feesom}{cost dummy variable; 1 if an annual user fee is paid at Lake somerville}
    \item{costcon}{travel cost to Lake Conroe}
    \item{costsom}{travel cost to Lake Somerville}
    \item{costhoust}{travel cost to Lake Houston}
  }
}
\source{Journal of Business, Economics and Statistics}
\references{
Seller, Christine; Stoll, John R. and Jean--Paul Chavas (1985) \dQuote{Validation of Empirical Measures of Welfare Change: a Comparison of Nonmarket Techniques}, \emph{Land Economics}, \bold{61(2)}, 156--175.

Gurmu, Shiferaw and Pravin K. Trivedi (1996) \dQuote{Excess Zeros in Count Models for Recreational Trips}, \emph{Journal of Business and Economics Statistics}, \bold{14(4)}, 469--477.

Santos Silva, Jao M. C. (2001) \dQuote{A Score Test for Non--nested Hypotheses with Applications to Discrete Data Models}, \emph{Journal of Applied Econometrics}, \bold{16(5)}, 577--597, \doi{10.1002/jae.601}
.

}
\keyword{datasets}
