\name{urban_crimes}
\docType{data}
\alias{urban_crimes}
\title{Urban Crimes in Pitsburg}
\description{monthly observations of 138 census tracts from 2008-01 to 2013-12

    \emph{number of observations} : 9936

    \emph{number of  time-series } :  72 

    \emph{country} :  Pitsburg 

    \emph{package} : 

}
\usage{data(urban_crimes)}
\format{A dataframe containing:
  \describe{
    \item{id}{the id of the tract}
    \item{month}{the month}
    \item{part1}{number of Part I offenses}
    \item{part2}{number of Part II offenses}
    \item{pop}{total population}
    \item{density}{population density per square mile}
    \item{medinc}{median income}
    \item{dropout}{dropout rate age 16-19}
    \item{unemp}{civilian unemployment rate}
    \item{poverty}{poverty rate}
    \item{under18}{percentage of total population under 18}
    \item{gpquarter}{group quarter proportion}
    \item{black}{percentage of total population that is African-American}
    \item{lesshsd}{percentage of population (age 25 and over) with less than a high school degree}
    \item{bachelor}{percentage of population (age 25 and over) with a bachelor degree or higher}
    \item{rental}{rental housing units as a percentage of occupied housing units}
    \item{samehouse}{percentage of households having been in the same house for more than one year}
    \item{female}{percentage of female headed household}
    \item{vacancy}{housing units vacancy rate}
  }
}
\source{JAE data archive}
\references{
Liesenfeld, Roman; Richard, Jean-Francois and Jan Vogler (2017) \dQuote{Likelihood Based Inference and Prediction in Spatio-temporal Panel Count Models for Urban Crimes}, \emph{Journal of Applied Econometrics}, \bold{32(3)}, 600-620.

}
\keyword{datasets}
